Close scanners

+++ rmi-java/src/clienteserver/Email.java
@@ -7,16 +7,16 @@ public class Email implements Serializable{
 	 * 
 	 */
 	private static final long serialVersionUID = 1L;
-	private double id;
+	private long id;
 	private String sender;
 	private String recipient;
 	private String subject;
 	private String body;
 	
-	public double getId() {
+	public long getId() {
 		return id;
 	}
-	public void setId(double id) {
+	public void setId(long id) {
 		this.id = id;
 	}
 	public String getSender() {
